feat: created one styles file per component
