+++ src/components/Analytics/styles.js
@@ -1,0 +1,24 @@
+import {makeStyles } from "@material-ui/core/styles";
+import * as COLORS from "../../constants/colors";
+
+export const useStyles = makeStyles((theme) => ({
+  root: {
+    "& .country--cell": {
+      backgroundColor: COLORS.COUNTRY,
+      border: "2px solid white",
+    },
+    "& .confirmed--cell": {
+      backgroundColor: COLORS.CONFIRMED,
+      border: "2px solid white",
+    },
+
+    "& .recovered--cell": {
+      backgroundColor: COLORS.RECOVERED,
+      border: "2px solid white",
+    },
+    "& .deaths--cell": {
+      backgroundColor: COLORS.DEATHS,
+      border: "2px solid white",
+    },
+  },
+}));
