+++ src/components/CreateNews/styles.js
@@ -1,0 +1,10 @@
+import {makeStyles } from "@material-ui/core/styles";
+
+export const useStyles = makeStyles((theme) => ({
+  paper: {
+    marginTop: theme.spacing(8),
+    display: "flex",
+    flexDirection: "column",
+    alignItems: "center",
+  }
+}));
